Add -l and -i as short options to git push.
They represent --force-with-lease and --force-if-includes, respectively.
This patch also fixes an unrelated typo.

The idea behind this patch that users can type, for example, git push -il
instead of git push -f, and benefit from additional safety guarantees
while still enjoying the convenience of short flags.

Of course, a user could set up an alias for this himself, but you could
say that about any short option, as far as I can tell.

This commit does not change the fact that the semantics of most forms of
--force-with-lease are experimental and subject to change. So is -l.

Originally I had this as three commits, but then I realized combining
them would be more efficient in this case. It's quite a simple change.
Also, the changes are not automatically tested, due to the git testing
configuration mysteriously failing on me; so, I'm hoping the changes are
obviously right or wrong. They seem to work for what I've tried them by
hand for!

Here is everything the commit does, three small changes:

add -l as a short form of --force-with-lease, also updating the Documentation

add -i as a short form of --force-if-includes, also updating the Documentation

fix a random typo I happened to notice while running tests ('passin'->'passing')

Signed-off-by: wyattscarpenter <wyattscarpenter@gmail.com>